Make ruby lsp stuff work in all use cases
This update ensures that ruby-lsp will work correctly for the various
directory structures I use. Some org repos have a shared overcommit
setup, which I have a static config set up for and point to that if the
buffer is under a directory that also has a `.overcommit` directory.
Otherwise, if there is a Gemfile in the root, bundle with that,
otherwise just use the global lsp.

Once I got ruby-lsp working, the go-to definition functionality was only
partially supported, so I've also added solargraph with diagnostics
disabled so it can handle just go-to definition functionality.

In order to fix a warning about mismatched encodings when both clients
are loaded on a buffer, I've added an override to force both clients to
use "utf-16" as the encoding.